# docs/swagger.yaml
basePath: /cs/v1
definitions:
...
  course-service_internal_domain.AttemptAnswer:
    properties:
      attempt_question_id:
        type: integer
      id:
        type: integer
      selected_answer_id:
        type: integer
      text_response:
        type: string
    type: object
  course-service_internal_domain.AttemptPagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  course-service_internal_domain.AttemptQuestionState:
    properties:
      answers:
        items:
          $ref: '#/definitions/course-service_internal_domain.PublicQuestionAnswer'
        type: array
      attempt_id:
        type: integer
      id:
        type: integer
      instructor_comment:
        type: string
      order_index:
        type: integer
      points:
        type: integer
      question:
        $ref: '#/definitions/course-service_internal_domain.Question'
      question_id:
        type: integer
      response:
        items:
          $ref: '#/definitions/course-service_internal_domain.AttemptAnswer'
        type: array
      score_awarded:
        type: integer
      text:
        type: string
      type:
        $ref: '#/definitions/course-service_internal_domain.TestingQuestionType'
    type: object
  course-service_internal_domain.AttemptState:
    properties:
      attempt:
        $ref: '#/definitions/course-service_internal_domain.TestAttempt'
      pagination:
        $ref: '#/definitions/course-service_internal_domain.AttemptPagination'
      questions:
        items:
          $ref: '#/definitions/course-service_internal_domain.AttemptQuestionState'
        type: array
      remaining_seconds:
        type: integer
    type: object
...
  course-service_internal_domain.PublicQuestionAnswer:
    properties:
      id:
        type: integer
      question_id:
        type: integer
      text:
        type: string
    type: object
  course-service_internal_domain.Question:
    properties:
      answers:
        items:
          $ref: '#/definitions/course-service_internal_domain.QuestionAnswer'
        type: array
      bank_id:
        type: integer
      difficulty_level:
        type: integer
      id:
        type: integer
      points:
        type: integer
      text:
        type: string
      type:
        $ref: '#/definitions/course-service_internal_domain.TestingQuestionType'
    type: object
  course-service_internal_domain.QuestionAnswer:
    properties:
      id:
        type: integer
      is_correct:
        type: boolean
      question_id:
        type: integer
      text:
        type: string
    type: object
...
  course-service_internal_domain.Test:
    properties:
      available_from:
        type: string
      available_to:
        type: string
      bank_ids:
        items:
          type: integer
        type: array
      course_id:
        type: integer
      course_section_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      duration_seconds:
        type: integer
      generation_settings:
        items:
          type: integer
        type: array
      id:
        type: integer
      max_attempts:
        type: integer
      max_score:
        type: integer
      questions_count:
        type: integer
      title:
        type: string
    type: object
  course-service_internal_domain.TestAttempt:
    properties:
      attempt_number:
        type: integer
      completed_at:
        type: string
      expires_at:
        type: string
      grade_status:
        $ref: '#/definitions/course-service_internal_domain.TestGradeStatus'
      id:
        type: integer
      started_at:
        type: string
      status:
        $ref: '#/definitions/course-service_internal_domain.TestAttemptStatus'
      test_id:
        type: integer
      total_score:
        type: integer
      user_id:
        type: integer
    type: object
  course-service_internal_domain.TestAttemptStatus:
    enum:
    - IN_PROGRESS
    - COMPLETED
    - NEEDS_GRADING
    type: string
    x-enum-varnames:
    - TestAttemptStatusInProgress
    - TestAttemptStatusCompleted
    - TestAttemptStatusNeedsGrading
  course-service_internal_domain.TestGradeStatus:
    enum:
    - NONE
    - READY
    - PRE_GRADED
    type: string
    x-enum-varnames:
    - TestGradeStatusNone
    - TestGradeStatusReady
    - TestGradeStatusPreGraded
  course-service_internal_domain.TestWithAttempts:
    properties:
      attempts:
        items:
          $ref: '#/definitions/course-service_internal_domain.TestAttempt'
        type: array
      test:
        $ref: '#/definitions/course-service_internal_domain.Test'
    type: object
  course-service_internal_domain.TestingQuestionType:
    enum:
    - SINGLE
    - MULTIPLE
    - TEXT
    type: string
    x-enum-varnames:
    - TestingQuestionTypeSingle
    - TestingQuestionTypeMultiple
    - TestingQuestionTypeText
...
  course-service_internal_transport_http_v1_request.CreateTest:
    properties:
      available_from:
        type: string
      available_to:
        type: string
      bank_ids:
        items:
          type: integer
        type: array
      course_id:
        type: integer
      course_section_id:
        type: integer
      description:
        type: string
      duration_seconds:
        type: integer
      generation_settings:
        items:
          type: integer
        type: array
      max_attempts:
        type: integer
      max_score:
        type: integer
      questions_count:
        type: integer
      title:
        type: string
    type: object
...
  course-service_internal_transport_http_v1_request.GradeAttempt:
    properties:
      scores:
        items:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.GradeAttemptQuestion'
        type: array
    type: object
  course-service_internal_transport_http_v1_request.GradeAttemptQuestion:
    properties:
      attempt_question_id:
        type: integer
      instructor_comment:
        type: string
      score_awarded:
        type: integer
    type: object
  course-service_internal_transport_http_v1_request.SaveAttemptAnswer:
    properties:
      attempt_question_id:
        type: integer
      selected_answer_id:
        type: integer
      selected_answer_ids:
        items:
          type: integer
        type: array
      text_response:
        type: string
    type: object
...
  course-service_internal_transport_http_v1_request.StartAttempt:
    properties:
      test_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  course-service_internal_transport_http_v1_request.UpdateTest:
    properties:
      available_from:
        type: string
      available_to:
        type: string
      bank_ids:
        items:
          type: integer
        type: array
      course_id:
        type: integer
      course_section_id:
        type: integer
      description:
        type: string
      duration_seconds:
        type: integer
      generation_settings:
        items:
          type: integer
        type: array
      id:
        type: integer
      max_attempts:
        type: integer
      max_score:
        type: integer
      questions_count:
        type: integer
      title:
        type: string
    type: object
  internal_transport_http_v1.ErrorResponse:
    properties:
      message:
        example: bad request
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for courses, sections, assignments, question banks and tests.
    Tests are also available under `/api` (same handlers as `/cs/v1`).
  title: Course Service API
  version: "1.0"
paths:
...
  /attempts:
    get:
      parameters:
      - description: Filter by test ID
        in: query
        name: test_id
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course-service_internal_domain.TestAttempt'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List attempts
      tags:
      - attempts
    post:
      consumes:
      - application/json
      parameters:
      - description: Questions page size
        in: query
        name: limit
        type: integer
      - description: Questions page offset
        in: query
        name: offset
        type: integer
      - description: Attempt
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.StartAttempt'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course-service_internal_domain.AttemptState'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Start test attempt
      tags:
      - attempts
  /attempts/{attempt_id}:
    get:
      parameters:
      - description: Attempt ID
        in: path
        name: attempt_id
        required: true
        type: integer
      - description: Questions page size
        in: query
        name: limit
        type: integer
      - description: Questions page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.AttemptState'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get attempt state
      tags:
      - attempts
  /attempts/{attempt_id}/answers:
    put:
      consumes:
      - application/json
      parameters:
      - description: Attempt ID
        in: path
        name: attempt_id
        required: true
        type: integer
      - description: Answer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.SaveAttemptAnswer'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Save attempt answer
      tags:
      - attempts
  /attempts/{attempt_id}/submit:
    post:
      parameters:
      - description: Attempt ID
        in: path
        name: attempt_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.TestAttempt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Submit attempt for grading
      tags:
      - attempts
...
  /tests:
    delete:
      parameters:
      - description: Test ID (query)
        in: query
        name: test_id
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Delete test
      tags:
      - tests
    get:
      parameters:
      - description: Filter by course ID
        in: query
        name: course_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course-service_internal_domain.Test'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List tests
      tags:
      - tests
    post:
      consumes:
      - application/json
      parameters:
      - description: Test
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.CreateTest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/course-service_internal_domain.Test'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Create test
      tags:
      - tests
    put:
      consumes:
      - application/json
      parameters:
      - description: Test
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateTest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Test'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Update test
      tags:
      - tests
  /tests/{test_id}:
    delete:
      parameters:
      - description: Test ID (query)
        in: query
        name: test_id
        type: integer
      - description: Test ID (path)
        in: path
        name: test_id
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Delete test
      tags:
      - tests
    get:
      parameters:
      - description: Test ID
        in: path
        name: test_id
        required: true
        type: integer
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.TestWithAttempts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get test info with user attempts
      tags:
      - tests
    put:
      consumes:
      - application/json
      parameters:
      - description: Test ID (path)
        in: path
        name: test_id
        type: integer
      - description: Test
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.UpdateTest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.Test'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Update test
      tags:
      - tests
  /tests/{test_id}/attempts:
    get:
      parameters:
      - description: Test ID
        in: path
        name: test_id
        required: true
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/course-service_internal_domain.TestAttempt'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: List attempts for a test
      tags:
      - tests
  /tests/{test_id}/attempts/{attempt_id}:
    get:
      parameters:
      - description: Test ID
        in: path
        name: test_id
        required: true
        type: integer
      - description: Attempt ID
        in: path
        name: attempt_id
        required: true
        type: integer
      - description: Questions page size
        in: query
        name: limit
        type: integer
      - description: Questions page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.AttemptState'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Get attempt state within a test
      tags:
      - tests
  /tests/{test_id}/attempts/{attempt_id}/grades:
    post:
      consumes:
      - application/json
      parameters:
      - description: Test ID
        in: path
        name: test_id
        required: true
        type: integer
      - description: Attempt ID
        in: path
        name: attempt_id
        required: true
        type: integer
      - description: Scores
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/course-service_internal_transport_http_v1_request.GradeAttempt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course-service_internal_domain.TestAttempt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_transport_http_v1.ErrorResponse'
      summary: Grade attempt manually
      tags:
      - tests
schemes:
- http
swagger: "2.0"